src/generation/requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.8
decorator==5.1.1
diffusers==0.31.0
dill==0.3.8
docker-pycreds==0.4.0
einops
eval-type-backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
gitpython==3.1.43
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jinja2==3.1.4
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.7
lightning
markupsafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt-toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rootutils==1.0.7
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.18.0
setproctitle==1.3.4
setuptools==75.5.0
simsimd==6.1.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
stringzilla==3.10.10
sympy==1.13.1
thop
tokenizers==0.20.3
torch==2.5.1
torchshow==0.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wand==0.6.13
wandb==0.18.7
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.17.2
zipp==3.21.0
